use queues to mange buffer pointers shared between the data thread that gets data from the Alazar card and the application
Two queues are used: the  bufferQ and the dataQ.  As buffer memory is allocated
to the Alazar using  the AlazarPostAsyncBuffer API call, the buffer pointers are
pushed onto the bufferQ.  To get new data, the data thread pops the next buffer
off of the bufferQ and passes it to the Alazar dll calling
AlazarWaitAsyncBufferComplete.  When enough triggers have occurred to fill the
buffer, AlazarWaitAsyncBufferComplete returns and the buffer pointer is placed
onto the dataQ.  The application thread polls the dataQ for new data. When data
is available, the buffer pointer is popped off the dataQ.  When the application
is done processing the data buffer, the pointer is pushed back onto the bufferQ.